fix potential memory overflow on gettime functions use better code for elapsed time

+++ advanced_functions.h
@@ -28,8 +28,8 @@ extern int check_for_script_file(const char* ors_boot_script);
 extern int run_ors_script(const char* ors_script);
 
 // general system functions
-long timenow_usec(void);
-long timenow_msec(void);
+long long timenow_usec(void);
+long long timenow_msec(void);
 char* readlink_device_blk(const char* Path);
 unsigned long Get_File_Size(const char* Path);
 unsigned long long Get_Folder_Size(const char* Path);
